Create separate header and source files for gen alpha time integrator
