+++ src/rigid_pendulum_poc/CMakeLists.txt
@@ -3,6 +3,7 @@ target_sources(${oturb_lib_name}
     #C++
     # diagnostics.cpp
     # io.cpp
+    generalized_alpha_time_integrator.cpp
     quaternion.cpp
     solver.cpp
     state.cpp
